requirements.txt
albumentations==1.0.3
imageio==2.9.0
numpy==1.19.2
Pillow==9.0.1
pynvml==11.4.1
tensorboardX==2.5
torch==1.4.0
torchvision==0.5.0
tqdm==4.49.0
yacs==0.1.8